Fix bug that causes `npm init` to silently fail.
**What I did**

1. create a new project directory && `cd` into it
1. create a new node_modules subdirectory (no longer necessary, but old habits die hard)
1. run `npm init` to create a pristine package.json

**What I expected**

To walk through a series of prompts to fill out a new package.json file and for that new file to get written to the current directory. Or at least an error if there was a problem.

**What actually happened**

After submitting a response to the "entry point" prompt, `npm init` silently exits with no error and no package.json created.