fix: accounts dot vs dash version/license config
Fixes a compat problem with the new `@npmcli/config` module that changes
the way these properties are handled, these aliases are no longer
auto-merged and instead live as two different properties in the new
config scheme. With that in mind this adds an extra check that validates
that the value retrieved by dot-separated property is actually different
from the default value defined in the global npm config before trying to
use that - this way users defining dash-separated configs can actually
use their values.